+++ testData/editor/highlighting/syntaxHighlighter/backquote.bash
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
 
-#The backquote characters are lexer highlights and not handled by the annotator
-`<info textAttributesKey="BASH.INTERNAL_COMMAND">echo</info> <info textAttributesKey="BASH.BACKQUOTE">hello</info> $var`
+# The backquote characters are lexer highlights and not handled by the annotator
+# $var is highlighted by the syntax highlighter, not by the annotator
+`<info textAttributesKey="BASH.INTERNAL_COMMAND">echo</info> hello $var`